Switch to new guacamole Docker image and remove nginx Docker image
The new guacamole image supports HTTP header auth, which has been
enabled.  In lieu of the nginx Docker container, we will use a GSSAPI-
and PAM-aware Apache httpd installation as our remote proxy.  This way
we can authenticate via Kerberos and leverage pam-sss to take
advantage of the HBAC rules in FreeIPA.

It makes sense to install httpd directly on the host system, since it
is using the host keytab as well as the host PAM and sssd
configurations.  It is probably possible to bind mount enough files
and directories to make it all work inside a Docker container, but a
direct installation is by far the simpler route.

Also remove unnecessary quotes and unnecessary networks and links
blocks in the docker-compose.yml file.  docker-compose creates a
network for this composition by default, and for this reason links are
deprecated.