Use PostgreSQL 13
The version of the JDBC PostgreSQL driver currently used in the
official Guacamole Docker image is too old to work with PostgreSQL 14
or later.  It produces errors that look like this:

[http-nio-8080-exec-5] WARN  o.a.g.e.AuthenticationProviderFacade - The "postgresql" authentication provider has encountered an internal error which will halt the authentication process. If this is unexpected or you are the developer of this authentication provider, you may wish to enable debug-level logging. If this is expected and you wish to ignore such failures in the future, please set "skip-if-unavailable: postgresql" within your guacamole.properties.
20:41:25.627 [http-nio-8080-exec-5] ERROR o.a.g.rest.RESTExceptionMapper - Unexpected internal error:
Error querying database.  Cause: org.postgresql.util.PSQLException: The authentication type 10 is not supported. Check that you have configured the pg_hba.conf file to include the client's IP address or subnet, and that it is using an authentication scheme supported by the driver.
The error may exist in org/apache/guacamole/auth/jdbc/user/UserMapper.xml
The error may involve org.apache.guacamole.auth.jdbc.user.UserMapper.selectOne
The error occurred while executing a query
Cause: org.postgresql.util.PSQLException: The authentication type 10 is not supported. Check that you have configured the pg_hba.conf file to include the client's IP address or subnet, and that it is using an authentication scheme supported by the driver.