Make dbinit a named volume instead of a bind mount
The 1.3+ versions of the Docker images guacamole/guacamole and
guacamole/guacd run as a non-root UID, so they are unable to copy
files into the bind mount directory since it is automatically created
with the default umask of 0022.  Using a named volume gets around this
limitation.

Making use of this syntax requires upgrading the composition's
version.